Relocate `Resource` to `org.junit.platform.commons.io` package (#4891)
This PR introduces a new `Resource` interface in the new
`org.junit.platform.commons.io` package.

The old interface extends the new one so it can be used in places that 
now consume the new interface.

The resource finding/scanning methods in `ReflectionSupport` are now 
deprecated in favor of the new ones in the `ResourceSupport` class.

Since `Predicate<...io.Resource>` and `Predicate<...support.Resource>`
have the same erasure, this PR introduces ResourceFilter (mimicking the 
existing `ClassFilter`) in order to be able to overload methods such as 
`addResourceContainerSelectorResolver` in
`EngineDiscoveryRequestResolver.Builder`.

Resolves #4885.